requirements.txt
discord.py==2.3.2
google-generativeai==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.1.0
python-dotenv==1.0.1
apscheduler==3.10.4
pytz==2024.1
aiohttp==3.9.5
oauth2client==4.1.3
Pillow==11.3.0
audioop-lts==0.2.1
PyNaCl==1.5.0